requirements.txt
fastapi
uvicorn
gunicorn; platform_system != "Windows"         # gunicorn is for Linux/macOS, avoid on Windows

python-multipart
PyYAML>=6.0
python-dotenv>=1.0.0

crewai==1.8.1
crewai_tools>=0.12.0

openai>=1.0.0
litellm<1.60.0
groq>=0.13.1

# Avoid pysqlite3-binary on Windows; keep it for non-Windows (optional)
pysqlite3-binary==0.5.4; platform_system != "Windows" and python_version < "3.12"

pydantic>=2.5.0